Stop the '</'> and '[/'] marks from intertwining
The function evil-visual-paste would assign to evil-visual-point/-mark
the same markers used by the '[ and '] marks. Therefore, after a
single visual paste Normal mode "gv" would act funkily in that buffer
ever after. To reproduce:

* Enter a new buffer with the contents:

      x
      y

* With the cursor on "x", type "ylvpjxgv".

The expectation is that "x" should be re-selected. Instead the empty
lower line is selected due to it being the previously changed text.

This commit fixes this, and also removes unused definitions related to
evil-visual-previous-point/-mark.