Track composite changes to current insertion text
The implementation of evil-track-last-insertion introduced in commit
477fe8e83e58739215ea3b601098ee134484114f assumed that all relevant
changes look like atomic insertions/deletions, but e.g.
remove-yank-excluded-properties upon yanking causes a single
substitution change. This commit fixes that.

A change not centered around the end of the previous insertion is also
made to start a new range, like in Vim.

Fixes #1759