Fix evil-test-logical-visual-next-line in GUI
The test was passing when run on text terminals but not on graphical
terminals. This was because in TUIs, wrapping is indicated by a
backslash that counts toward the width returned by window-width, but in
GUIs a bent arrow (↩) is placed in the fringe instead, so one more
character was needed in order to display a continuation line.