add keep-active-mark property to states
This property is used to determine whether we should enter visual state
when the mark is activated. This is useful for states like emacs state
where we want to keep the mark active but not enter visual state. This
change enables users to customize this behavior for their own states.