Reselect minibuffer when quitting command-line win
Pressing C-f while editing an Ex command to open the command-line
window and later quitting that window used to not reselect the Ex
command line, but instead just continue the recursive edit in some
previous window. This was specially confusing before commit
02004bce6884619e687654d333b75c90f8fc27d0, as the evil-ex-abort hack
did not trigger, leaving you in a bad state with
evil-ex-current-buffer set to t.

This commit fixes that by setting the delete-window window parameter
on the command-line window to a function that switches back to the
minibuffer after deleting the window, which closes #1379.

This commit also replaces the split-window call with a display-buffer
invocation, to always spawn the window at the bottom of the frame, but
also allow the user to override this behavior.