Added clarification to evil-define-key for escaping modifier keys
It is useful to have clarification in the doc string for
evil-define-key that it is necessary to escape modifiers keys for
users that are accustomed to using functions like global-set-key or
keymap-global-set where this is not necessary. This is also not made
particularly clear in the define-key function itself, which is a
legacy function.