Fix issue: No visible text when cursor is moved to the end of the buffer
When the user sets `scroll-conservatively` to a value greater than 100,
moving the cursor to `(point-max)` with `(evil-goto-line nil)` can cause
all text to be positioned above the window start, making it invisible to
the user. This issue is more likely to occur after scaling the text
using functions such as `text-scale-increase` or `text-scale-set`.

This commit resolves this issue by adjusting the window's view using
`(recenter -1)`. By doing so, it prevents the cursor from being placed
off-screen, ensuring that the user can always see the relevant text when
moving to the end of the buffer.

When `scroll-conservatively` is less than or equal to 100, Emacs
recenters the screen using `(recenter nil)`, which recenters with point
on the middle line. This behavior does not match Vim's. To align with
Vim's behavior, we use `(recenter -1)`, regardless of the value of
`scroll-conservatively`.