Fix touch screen gestures (e.g., scrolling) for Emacs 30
Emacs 30 introduces improved touch-screen support (via the Android port), see
“Touchscreen Events” in the Emacs Lisp reference manual: By default, touch
screen gestures are not processed when a command is bound to `down-mouse-1` (as
`evil-mouse-drag-region` is), as that is taken as an indication that touch input
should be translated to `down-mouse-1` followed by mouse motion events.

This change applies the `ignored-mouse-command` property to
`evil-mouse-drag-region` that is also present on `mouse-drag-region` in Emacs
30.  This has the effect that touch screen gestures are prioritized and mouse
events are only emitted when no gesture was detected.