Remove evil--ex-expression
As after-change-functions such as evil--ex-update are not guaranteed
to run, this commit ensures evil-ex re-parses the final Ex command
string instead of relying on evil--ex-expression. This makes the
workaround in commit 6afd86bbc740f3008e91881f990bf346b31d3f0b (Fix for
evil-ex-teardown, 2025-01-03) to not overwrite a shortcut in
evil--ex-expression no longer necessary.

Closes #1949